+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to viky.ai will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## 2019-01
+
+### Added
+
+- NLP: add Javascript variable called `language` corresponding to detected language which can be used in solution build. [#62](https://github.com/viky-ai/viky-ai/pull/62)
+
+
 ## 2019-12
 
 ### Added